create clipping planes

+++ src/bim-viewer-wiv.js
@@ -11,13 +11,24 @@ import {
   labeling,
 } from "../modules/twin.js";
 
+import {
+  IFCWALL,
+  IFCWALLSTANDARDCASE,
+  IFCSLAB,
+  IFCWINDOW,
+  IFCMEMBER,
+  IFCPLATE,
+  IFCCURTAINWALL,
+  IFCDOOR,
+} from "web-ifc";
+
 import Stats from "stats.js/src/Stats";
 
 // Get the URL parameter
 const currentURL = window.location.href;
 const url = new URL(currentURL);
 const currentModelId = url.searchParams.get("id");
-const toggle = {}
+const toggle = {};
 
 // Get user
 let currentUser = "";
@@ -108,59 +119,67 @@ async function loadIfc(ifcURL) {
   toggle.plans = false;
   const plansMenu = document.getElementById("plans-menu");
   toggleVisibility(plansButton, toggle.plans, plansMenu);
-  
+
   // Toggle left menu ⬅️
   document.getElementById("toolbar").onclick = () => {
-  let plans = !document.getElementById("plans-menu").classList.contains("hidden");
-  let ifc = !document.getElementById("ifc-tree-menu").classList.contains("hidden");
-  toggle.left = plans || ifc;
-  toggle.left ? document.getElementById("left-menu").classList.remove('hidden') :
-  document.getElementById("left-menu").classList.add('hidden')
-  }
-  
-
+    let plans = !document
+      .getElementById("plans-menu")
+      .classList.contains("hidden");
+    let ifc = !document
+      .getElementById("ifc-tree-menu")
+      .classList.contains("hidden");
+    toggle.left = plans || ifc;
+    toggle.left
+      ? document.getElementById("left-menu").classList.remove("hidden")
+      : document.getElementById("left-menu").classList.add("hidden");
+  };
 
   await viewer.plans.computeAllPlanViews(model.modelID);
 
-  const lineMaterial = new LineBasicMaterial({color: 'black'});
+  const lineMaterial = new LineBasicMaterial({ color: "black" });
   const baseMaterial = new MeshBasicMaterial({
     polygonOffset: true,
-    polygonOffsetFactor:1,
-    polygonOffsetUnits: 1
+    polygonOffsetFactor: 1,
+    polygonOffsetUnits: 1,
   });
 
-  await viewer.edges.create('plan-edges', model.modelID, lineMaterial, baseMaterial);
+  await viewer.edges.create(
+    "plan-edges",
+    model.modelID,
+    lineMaterial,
+    baseMaterial
+  );
 
   // Floor plan viewing
   const allPlans = viewer.plans.getAll(model.modelID);
   const plansContainer = document.getElementById("plans-menu");
 
-  for (const plan of allPlans){
-    const currentPlan = viewer.plans.planLists[model.modelID][plan]
-    if ((currentPlan.name).includes("LV")){
-    const planButton = document.createElement('button');
-    planButton.classList.add('levels')
-    plansContainer.appendChild(planButton);
-    planButton.textContent = currentPlan.name;
-    planButton.onclick = () => {
-      viewer.plans.goTo(model.modelID, plan, true)
-      viewer.edges.toggle('plan-edges', true)
-      togglePostproduction(false);
-      toggleShadow(false);
+  for (const plan of allPlans) {
+    const currentPlan = viewer.plans.planLists[model.modelID][plan];
+    if (currentPlan.name.includes("LV")) {
+      const planButton = document.createElement("button");
+      planButton.classList.add("levels");
+      plansContainer.appendChild(planButton);
+      planButton.textContent = currentPlan.name;
+      planButton.onclick = () => {
+        viewer.plans.goTo(model.modelID, plan, true);
+        viewer.edges.toggle("plan-edges", true);
+        togglePostproduction(false);
+        toggleShadow(false);
+      };
     }
   }
-  }
 
-  const button = document.createElement('button');
+  const button = document.createElement("button");
   plansContainer.appendChild(button);
-  button.classList.add('button')
-  button.textContent = 'Exit Level View';
+  button.classList.add("button");
+  button.textContent = "Exit Level View";
   button.onclick = () => {
     viewer.plans.exitPlanView();
-    viewer.edges.toggle('plan-edges', false)
+    viewer.edges.toggle("plan-edges", false);
     togglePostproduction(true);
     toggleShadow(true);
-  }
+  };
 
   viewer.IFC.getSpatialStructure(model.modelID);
   await viewer.shadowDropper.renderShadow(model.modelID);
@@ -185,13 +204,30 @@ dimensionsButton.onclick = () => {
   let visibility = toggle.dimensions ? "Hide" : "Show";
   let button = document.getElementById("dimensions");
   button.setAttribute("title", `${visibility} ${button.id}`);
-  toggle.dimensions? button.classList.add("selected-button") : button.classList.remove("selected-button");
+  toggle.dimensions
+    ? button.classList.add("selected-button")
+    : button.classList.remove("selected-button");
   clicked = 0;
 };
 
-// Double click → Dimensions
+// Clipping planes
+const clippingButton = document.getElementById("clipping");
+toggle.clipping = false;
+clippingButton.onclick = () => {
+  toggle.clipping = !toggle.clipping;
+  viewer.clipper.active = toggle.clipping
+  let visibility = toggle.clipping ? "Hide" : "Show";
+  let button = document.getElementById("clipping");
+  button.setAttribute("title", `${visibility} ${button.id}`);
+  toggle.clipping
+    ? button.classList.add("selected-button")
+    : button.classList.remove("selected-button");
+}
+
+
+// Click → Dimensions
 window.onclick = () => {
-  if (clicked > 0) {
+  if (clicked > 0 && toggle.dimensions) {
     viewer.dimensions.create();
   }
   clicked++;
@@ -204,12 +240,16 @@ window.onkeydown = (event) => {
     viewer.IFC.selector.unHighlightIfcItems();
   }
   if (event.code === "Space") {
-    viewer.context.fitToFrame()
+    viewer.context.fitToFrame();
   }
 
-  if (event.code === "Delete") {
+  if (event.code === "Delete" && toggle.dimensions) {
     viewer.dimensions.delete();
   }
+  if (event.code === "Delete" && toggle.clipping) {
+    viewer.clipper.deletePlane();
+  }
+
 };
 
 // Properties 📃📃📃📃📃📃📃📃📃📃📃📃📃📃📃📃📃📃📃
@@ -226,7 +266,13 @@ window.ondblclick = async () => {
   if (!result) return;
   const { modelID, id } = result;
   const props = await viewer.IFC.getProperties(modelID, id, true, false);
-  createPropsMenu(props)
+
+  createPropsMenu(props);
+
+  if(toggle.clipping){
+    viewer.clipper.createPlane();
+  }
+
 };
 
 function createPropsMenu(props) {
@@ -359,13 +405,13 @@ window.oncontextmenu = () => {
   labeling(scene, collisionLocation, currentUser);
 };
 
-function toggleShadow(active){
+function toggleShadow(active) {
   const shadows = Object.values(viewer.shadowDropper.shadows);
   for (shadow of shadows) {
     shadow.root.visible = active;
   }
 }
 
-function togglePostproduction(active){
+function togglePostproduction(active) {
   viewer.context.renderer.postProduction.active = active;
 }
